app/utils.py
from typing import Union
from jose import jwt

from app.core.config import settings

def verify_password_reset_token(token: str) -> Union[str, None]:
    try:
        decoded_token = jwt.decode(token, settings.SECRET_KEY, algorithms=["HS256"])
        return decoded_token["email"]
    except jwt.JWTError:
        return None